Updated tests to be compatible with Travis-CI
One of the tests checks the form of the input data files obtained from EIA. Because these files are not in the online repository, this test will fail with an error and stop the test process. To avoid that error, the test will be automatically skipped when running on Travis.

In addition, the test now uses a setUp function to load the required data from the EIA data file.