Configure Travis-CI
Add the YAML file required to set up Travis-CI and the supporting requirements.txt file that includes any Python packages used in Scout that aren't part of base Python.

Note that the version of numpy indicated in the requirements.txt file was chosen based on the version used by Travis-CI, since using a newer version than that already available requires a few minutes of reinstallation each time a new commit is pushed to the remote repository and tested.