Added function to subset data from KDBOUT array
Since the KDBOUT data, and thus the array in which those data are stored, include energy use and several other parameters indexed by census division, building type, end use, and fuel type, a function was added to extract the data for a specified parameter from the larger array. A corresponding test was added as well. This new data subset function will be used by a larger function (forthcoming) that prepares data for each leaf node.