Added function to convert data into dict output
Added a function that converts the input data from EIA and the thermal load components study into output dicts for each commercial building leaf/terminal node in the microsegments JSON database. This function selects the appropriate data from the various arrays in which they are stored and combines the relevant extracted data, as needed, to obtain dicts with keys for each year of data and values for each year corresponding to the energy use for that microsegment. The newly created dicts will then be inserted at each leaf node. Companion tests are also included.