Added function to traverse JSON; fixed bugs
The primary revision is the addition of the 'walk' function, which traverses the microsegments JSON database recursively, and upon reaching a leaf/terminal node, calls a function that uses the list of keys that uniquely define that leaf node to identify, prepare, and insert the energy data for that node. Additionally, several bugs were fixed.

Bugs squashed:
- File name text pointing to service demand data
- Text string in demand_typedict for thermal load from equipment
- No longer producing data from the catg_data_selector function with the 'Year' column as integers, since the years are used as keys in dicts that eventually should be output as a JSON file, and valid JSON only has strings as keys
- Placeholder rows from the service demand data are now discarded and technologies that don't have a specific year in their description are passed through without regex filtering; both cases are now tested as well