Improved compatibility with commercial buildings
In anticipation of a microsegments JSON file that includes both residential and commercial building types, an if statement that checks to make sure the appropriate (residential or commercial) building type is in the list of keys before proceeding with data extraction was added to the 'walk_techdata' function.

Additionally, the procedure to import the census division to climate zone data file was moved inside the 'main' function to ensure that if mseg_techdata is called by another function, it does not silently trigger a file import action.

Also, a more detailed explanation of the purpose of the _Rev_ version of the census division to climate zone conversion data file was added to the comments.