Revised calculator appearance and corresponding JS
In preparation for adding commercial building data, the appearance and configuration of the calculator has been changed slightly to allow selection of only residential or commercial building types at one time. The addition of a new radio button to select either residential or commercial buildings to populate the building type selection field required some updates to the JavaScript/jQuery. Also, a typo in the JavaScript was corrected.