Updated envelope to include equipment heat gain
Previously, heat gain from equipment and people had been left out of the options for energy associated with the building envelope. People were excluded because those loads cannot be modified directly. Equipment heat gain was excluded presumably because it is not associated with the envelope components themselves. It is, however, useful information, and since the data are in the JSON database, the JavaScript has been modified to make the data available through the calculator interface.