Add realistic stocks and flows to master microsegment calculation
Stock and flow dynamics have been added to the ‘partition_microsegment’
function in ‘run.py’.  Such dynamics are needed to run measure adoption
scenarios beyond the ‘Technical potential,’ which assumes that all AEO
baseline stock is captured by an efficient measure in each year of the
modeling time horizon. In reality, only a portion of each year’s
baseline stock will be available for an efficient measure to capture,
given the need for equipment in new homes and replacement of baseline
equipment that has reached the end of its useful life.

The primary stocks and flows for an efficient measure across time are
as follows:

Stocks

Competed (efficient)
- New and replacement stock for the current year that is captured by
the efficient measure
Competed (baseline)
- New and replacement stock for the current year that is captured by
the comparable baseline technology
Non-competed (efficient)
- Existing stock that was captured in previous years by the efficient
measure
Non-competed (baseline)
- Existing stock that was captured in previous years by the baseline
technology

Flows*

New additions (efficient)
- Efficient stock associated with new homes added in the given year
New additions (baseline)
- Baseline stock associated with new homes added in the given year
Replacements (baseline -> baseline)
- Baseline stock at end of its life that goes to a newer version of the
conventional technology
Replacements (baseline -> efficient)
- Baseline stock at end of its life that changes to the efficient
measure
Replacements (efficient -> efficient)**
- Efficient stock at end of its life that goes to a newer version of
the efficient measure

Currently, the above stocks and flows are utilized in a ‘Max adoption
potential’ scenario where it is assumed that an efficient measure
captures all competed stock in a given year (e.g., ‘Competed
(baseline)’ above is zero). Longer term, an ‘Adjusted adoption
potential’ scenario may be added in which the competed stock is
apportioned between the baseline technology and an efficient measure,
given externally defined adoption modeling inputs (see
‘diffuse_eff_frac’ placeholder variable in ‘partition_microsegment’).

In running a Max adoption potential scenario on the set of previously
committed sample measures, a small flaw in the measure competition
routine was exposed.  Specifically, the competed market share of a
measure was being applied to its competed savings for a given year when
it should have been applied to its total savings in that year. The
‘compete_adjustment_dicts’ and ‘compete_adjustment’ functions have been
updated to fix this issue.

The ‘PartitionMicrosegmentTest’ in ‘run_test.py’ has been expanded to
include three ‘Max adoption potential’ test cases on top of the three
‘Technical potential’ scenarios that were already tested.

* Retrofits of existing stock are another possible flow, but are not
currently simulated. Given data on the % of the existing stock that is
retrofitted each year, this type of flow could be added in a future
commit.
**It is currently assumed that once adopted, efficient stock is always
updated with newer versions of the efficient measure and never changes
back to baseline stock.