Move to common data for all tests of com_mseg
Create a class that includes the definitions of all the major data structures with which the tests are run and switch all tests to be children of this common data setup class. This greatly simplifies the child classes and cleans up the test input and output data so that all cases are tested against all functions in com_mseg. Closes #89.