Added generation of error messages visible to users
Added jQuery code that checks the entries in the calculator form and generates explanatory error text if there are incomplete entries. The error text is added to the DOM just above the first calculator field. Clicking the update button triggers the entry checking process, clears any existing error text, and adds the appropriate warning text to the DOM. Other changes also ensure that the reported numbers are always zeroed out when errors are present. Closes #56.

During testing, it was revealed that when switching between water heating, drying, and cooking, once one had selected a fuel type for an end use, changing the end use would not require reselection of the fuel type before being able to get updated numbers because the previously selected fuel type was still being used. This problem was caused by reuse of the old value of the ft_only_sel variable, which is now (clumsily) fixed.

Also made some minor tweaks to the HTML for accessibility.