Updated commercial electricity prices to $/source MMBtu
In the original version of this file on the run_scheme branch (Cost_S-S_CO2.txt), the commercial electricity price had units of $/MMBtu site energy. These prices were converted to source energy to be consistent with the residential price data already in the file in commit #cffb85a. Though the calculator does not currently use these price data, for the sake of consistency, the same update to commercial electricity prices is now made here. These changes are in partial response to issue #96.