Added function to extract technology lifetime data
The primary change is the addition of the function and supporting test that extracts technology lifetime data, defined using the same parameters as in the residential mseg_techdata.py (e.g., ‘average’ and ‘range’), and converts it into a formatted dict for output to a JSON file.

Additionally, the helper function dict_check was moved to the top level CommonUnitTest class so that it would be available to all test sub-classes, and the content in the test script was reshuffled so that the tests and test data would be in the same order as the functions under test.