Check validity of candidate key chains for measure performance dict
This commit ensures that the key chains for all branches in the measure
performance dictionary are consistent with the applicable baseline
market that is defined for the measure.

In particular, if a candidate key chain for a measure’s performance
dictionary points to a set of technologies that cannot be found in the
‘technology’ attribute for the measure, that key chain path is removed
from the procedure that constructs the measure performance dictionary.

For example, consider a candidate measure performance key chain that
looks like this: [‘hot dry’, ‘assembly’, ‘natural gas’, ‘cooling’,
‘supply’]. Let’s say we follow this path down the levels of an input
dictionary containing all valid microsegments and find only one
applicable technology, ‘NGHP’ (natural gas heat pumps). If natural gas
heat pumps are not listed in the measure’s ‘technology’ attribute, this
key chain should not be included in the measure’s performance
dictionary.

To accommodate this additional check, a sample input dictionary
containing valid baseline microsegment information was added to the
existing ‘CreatePerformanceDictTest’ in ‘import_eplus_test.py’.