Modified cost_perf_extractor function and tests
The cost_perf_extractor function uses technology description text string matching to select the correct service demand data to use as weights with the technology cost and performance data. In the EIA AEO technology data, these strings are of various lengths, but in the service demand data, the strings are truncated to 44 characters. The cost_perf_extractor function was updated to similarly truncate the technology data string descriptions before attempting matching with the service demand data. To ensure the modifications were working as expected, an additional test case was added that requires that the technology description string be truncated.

Also, a docstring was added to the units_id function.