Adjust weighting of measure relative performance and market share
A measure’s overall relative performance and market share is now
weighted in each year and adoption scenario to properly account for the
relative influence of adoptions in the current year and all previous
years of the modeling time horizon.

The new relative performance weighting required a change in the
stock-and-flow assumptions of the Technical Potential (TP) scenario.
Previously, this scenario assumed that a measure replaces all
applicable baseline stock in each year it is on the market. This
approach was problematic in years where the efficiency of the baseline
stock dramatically improved over the previous year; such cases yielded
large decreases in the relative savings of the measure over the
baseline and associated jumps in stock-wide consumption for the TP over
time.

With this commit, a TP scenario now assumes that a measure replaces all
applicable baseline stock in its market entry year and subsequently
captures any new additions to the stock or replacements of previously
captured stock. Given this adjusted TP assumption, the relative savings
of captured new and replacement stock in each year can be weighed
against the relative savings of captured stock from all previous years;
this eliminates the previously observed year-to-year jumps in relative
savings and TP consumption.

Measure market shares are similarly weighted in the measure competition
routine to account for the relative savings contributions of the
captured stock in the current year vs. all previous years. For example,
assume the stock captured by a measure in a given year accounts for 10%
of its total savings in that year, while previously captured stock
accounts for 90% of these savings; also assume that the measure’s
calculated market share for the given year is 50%, while its combined
market share across all previous years is 30%. Given these assumptions,
the measure’s weighted market share for the current year is (50% * 10%
+ 30% * 90%)= 32%. This weighted market share can then be used to
adjust the measure’s uncompeted total savings in the current year.

To accommodate the above changes, several of the input/output values in
the tests of ‘run_test.py’ have been adjusted; however, the structure
of these tests remains the same as before. Minor improvements to code
comments and formatting have also been made throughout ‘run.py’.