Compete windows solar and windows conduction measures
This commit adjusts contributing microsegment information for measures
that affect the ‘windows solar’ and/or ‘windows conduction’ market
microsegments to indicate their applicability to a general ‘windows’
market. This ensures that windows solar and windows conduction measures
will be directly competed as part of the ‘adjust_savings’ function.
Direct competition of windows solar and conduction measures makes sense
in new and replacement decisions because both are tied to the same
window unit; thus, if a consumer decides to go with a measure that
affects window conduction, there will be one less unit for a measure
that affects window solar heat gain to capture, and vice versa.

This commit additionally multiplies stock square footage information by
one million to reflect the AEO reporting of this square footage
information in units of million square feet.

Small adjustments were made to the contributing microsegment key chain
definitions of the ‘ResCompeteTest’ in ‘run_test.py’ to reflect the new
naming of ‘window conduction’ and ‘window solar’ technologies as
‘window’ technologies for measure competition purposes. Additionally, a
test of the new ‘add_keyvals_restrict’ function was added to
‘run_test.py’; ‘add_keyvals_restrict’ operates much like the existing
‘add_keyvals’ function, but disallows the addition of lifetime
information for its two input dictionaries. The ‘add_keyvals_restrict’
function is used in cases where a measure has both ‘windows conduction’
and ‘windows solar’ contributing microsegment information that must be
combined into information for a single ‘windows’ contributing
microsegment. This windows microsegment is subsequently used in measure
competition routines.

Finally, stock and stock cost values in the
‘FindPartitionMasterMicrosegmentTest’ were multiplied by one million to
reflect the adjusted handling of AEO square footage input data.

Closes #119