Amend dynamic relative savings and IRR/payback calculations
This commit amends the method for adjusting a measure’s relative
savings percentage over time with a changing baseline performance level
(sample measure testing revealed the initial method yielded incorrect
values in certain cases). Values for ‘sample measure 13’ in the
‘FindPartitionMasterMicrosegmentTest’ of ‘run_test.py’ were adjusted
accordingly.

Additionally, the calculation of IRR and simple payback in the
‘metric_update’ function was restricted in cases where a measure’s
energy and carbon savings are zero. In these cases, the IRR/payback
calculation was yielding NaNs, when it was intended to have yielded a
999 (easier to handle in output plotting).