Add handling of secondary microsegment stocks-and-flows
In the commercial sector, Scout accounts for the secondary effects that
lighting efficiency measures have on heating and cooling energy use.
Such secondary heating and cooling microsegments require special
handling in stock-and-flow calculations, since associated energy/carbon
savings are not achieved through direct heating and cooling equipment
adoptions but instead indirectly through the adoption of new lighting
equipment.

Accordingly, stock-and-flow calculations for secondary heating and
cooling microsegments have been linked to the stocks-and-flows in their
associated primary lighting microsegments. This means that if only a
fraction of a primary lighting microsegment is competed and captured by
a lighting measure in each year, the same fraction of an associated
heating/cooling microsegment will be competed and captured. Primary and
secondary microsegments are associated by their climate zone, building
type, and structure type (e.g., ‘new’ or ‘existing’).

Information needed to link primary and secondary microsegments for
stock-and-flow adjustments is now stored in a new ‘secondary mseg
adjustments’ key in the existing ‘mseg_adjust’ measure attribute.

To accommodate the above changes, a new sample commercial lighting
measure has been added to the ‘FindPartitionMasterMicrosegmentTest’ in
‘run_test.py’. The test ensures that both the primary and secondary
energy/carbon impacts of the lighting measure are correctly calculated
over multiple years for a max adoption potential scenario where only a
fraction of the lighting stock is competed in each year. Additionally,
a few adjustments were made to the outcome values for existing sample
lighting measures to reflect the improved handling of secondary
microsegment stock-and-flow calculations.

Finally, the ‘PartitionMicrosegmentTest’ was changed to reflect the
addition of the ‘mseg_adjust’ dictionary as both an input and output of
the ‘partition_microsegment’ function being tested.

Closes #106