Add handling of secondary microsegment market share adjustments
In the commercial sector, Scout accounts for the secondary effects that
lighting efficiency measures have on heating and cooling energy use.
Such secondary heating and cooling microsegments require special
handling in market share calculations, since associated energy/carbon
savings are not achieved through direct heating and cooling equipment
adoptions but instead indirectly through the adoption of new lighting
equipment.

Accordingly, market share calculations for secondary heating and
cooling microsegments have been linked to the market shares calculated
for their associated primary lighting microsegments. This means that if
after a lighting measure is competed with other lighting measures it is
determined to capture only a fraction of a primary market microsegment
in each year, this captured fraction will be applied to all secondary
heating and cooling microsegments that are associated with the primary
microsegment. As mentioned in the previous commit, primary and
secondary microsegments are associated by their climate zone, building
type, and structure type (e.g., ‘new’ or ‘existing’).

Reflecting the approach above, the ‘adjust_savings’ function in
‘run.py’ has been enhanced to ensure that secondary microsegment
adjustments are properly accounted for. Specifically, when a primary
microsegment with secondary effects is competed in the ‘res_compete’ or
‘com_compete’ sub-functions of ‘adjust_savings,’ resultant market share
adjustments for that microsegment are stored in the measure’s
‘mseg_adjust’ attribute. Subsequently, when the associated secondary
microsegment is in line for an update, it is passed through a new
‘secondary_mktshare_adjust’ function, which adjusts its microsegment
values by the market shares previously calculated for the associated
primary microsegment.

To accommodate the above changes, a secondary cooling microsegment has
been added to one of the competing commercial lighting measures in
‘ComCompeteTest’ of ‘run_test.py’. The test verifies that this
secondary microsegment is properly adjusted in accordance with the
market share calculated for its associated primary lighting
microsegment. Additionally, ‘primary’ tags were added to the beginning
of several sample microsegments being run through the measure
competition tests; the microsegments had previously been missing these
tags, though this did not affect the validity of the tests.

Closes #94