Revise data_import function
The data_import function skips lines in the data file being read that do not have the same length as the defined dtype. The descriptions of some cooking equipment types include commas within their description. If the double-quoted string for the description isn't immediately preceded by a ',' with no spaces, the commas inside the string will be treated as delimiters as well. As a result, some cooking equipment types were not being included in the input data because the lines appeared to have too many columns. Adding the skipinitialspace option to csv.reader fixes this problem.