Fixed technology name matching problems
When selecting the service demand data using the technology name strings from the cost, performance, and lifetime data, matches were not always found.

Commit #2ea2bac partially addressed this problem for the case where the technology data weren’t being imported because there appeared to be too many columns in some lines of the data. There were a few other specific problems that were preventing matching and are addressed by this commit: the ampersand and double quote characters are represented as short text strings in the service demand data, and there were sometimes trailing spaces on the text string when it was truncated to 44 characters to match the maximum length of the technology names in the service demand data.

In addition to fixing failures to find matches, several functions were modified to support the production of an error message that reports any technology types that appear in the cost, performance, and lifetime data, but not in the service demand data. This error message is produced following the final call of the walk function, when the populated database structure is complete, and it produces to STDOUT the technology performance levels that did not match.

Tests were updated to support the additional outputs from the functions that were modified to report the non-matching technology names.