Added square footage data processing
Updated the tests and code to incorporate new and total square footage for all commercial buildings, specified by climate zone and building type. Functions that were modified to support the square footage data have updated docstrings that comply with Google’s Python style guide.

In the ‘json_interpreter’ function and the ‘CommercialDataSelectionTest’ test, lists were being modified in-place to overwrite or swap some entries in those lists with other entries. In both cases, the code was altered so that new lists are created and modified so that the original list retains its integrity and can be used for running other tests without causing unexpected errors or test failures.