Improve robustness of financial calculations/other minor adjustments
The checks for lists of all zeros in the financial metrics calculations
were not robust to cases where floating point subtractions of identical
numbers left very small differences (see
https://docs.python.org/2.7/library/decimal.html for more explanation).
To handle such cases, the numpy.isclose function is used to check for
approximately zero values (within a default tolerance of 10e-8).

Additionally, secondary market microsegment adjustment information was
added to the mseg_adjust attribute of packaged measures, and the
calculations for adjusting secondary market shares were slightly
modified to accommodate a packaged measure case.

Finally, secondary market microsegment and sub-market adjustment
information was added to several of the sample measures in
‘run_test.py’. The omission of this information does not affect most of
the tests, but the information should nevertheless be included for
completeness.