Revised structure of commercial energy data
The residential energy and stock data are recorded in a dict with the structure {'stock': {'2009': n, '2010': n2, … , '2040': n41}, 'energy': {'2009': n, '2010': n2, … , '2040': n41}} for each technology. The commercial data do not include total stock of each technology, thus after each technology key appeared the energy use data for each year. To simplify the code for measure evaluation, the 'stock' and 'energy' keys are now added to the commercial data. paralleling the structure for the residential data.