Migrate functions from 'run.py,' revise 'fill_measures' function
The function used to build a measure’s master microsegments in ‘run.py’
was moved to ‘measures_prep.py’ along with supporting functions and
tests; this function has been renamed from ‘mseg_find_partition’ to
‘fill_mkts’.

The ‘fill_mkts’ function finalizes measure markets as part of the
previously implemented ‘fill_measures’ routine. The latter has been
revised for this commit as follows:
•	The measure cost unit translation functions (‘translate_costs_res’
and ‘translate_costs_com’) have been moved to the ‘fill_mkts’ function.
•	A new approach is used to determine which input measures must be
finalized via ‘measures_prep’. Under this approach, a user flags the
measures needing updates by setting their ‘status->finalized’ attribute
to False. Measures that are not flagged for updating by the user but
have an incomplete ‘markets’ attribute are automatically added to the
list of measures to finalize; the user is warned about these added
measures.

To accommodate the above changes, the ‘FillMeasuresTest’ in
‘measures_prep_test.py’ has been adjusted to test: 1) an input measure
that is flagged for updating; 2) an input measure that does not need
updating; and 3) an input measure not flagged for updating that has
incomplete ‘markets’ information.

Finally, Google style Docstrings have been added for all classes and
functions in ‘measures_prep.py’ and ‘measures_prep_test.py’.