Account for market overlaps within packages
The primary change in this commit is the consideration and removal of
market overlaps between measures in a package as part of the
‘merge_measures’ function. Overlaps are removed via a
‘merge_contrib_msegs’ sub-function within ‘merge_measures’.

In removing market overlaps between measures in a package, the
‘merge_contrib_msegs’ function handles ‘full service’ and ‘add-on’
measure types differently.* While market microsegment stock, stock
cost, and lifetime data are evenly apportioned across the number of
overlapping full service measures, add-on measures assume the stock and
lifetime data of any full service measure(s) they overlap with, and
incrementally add to the stock cost of the overlapping full service
measure(s). Energy, carbon, and energy/carbon cost markets are evenly
apportioned across the total number of overlapping full service and add
on measures.

Additional minor changes for this commit include the following:
•	Rename ‘out_break_norm’ function to ‘div_keyvals’ and ‘reduce_sqft’
function to ‘div_keyvals_float_restrict,’ also adding a new
‘div_keyvals_float’ function for cases where all a dictionary’s
terminal leaf node values must be divided by a single number.
•	Adjust commercial building type naming conventions (e.g.,
‘healthcare’ changed to ‘health care’, ‘mercantile and service’ changed
to ‘mercantile/service’ to be consistent with ‘com_mseg.py’).
•	Remove ‘savings updated’ key from a measure’s ‘markets’ attribute (to
be handled in ‘run.py’).
•	Clarify comments and Docstrings throughout.

All relevant tests in ‘measures_prep_test.py’ have been adjusted to
accommodate the above changes. In particular, the MergeMeasuresTest now
merges multiple sample measures with overlapping markets, including
both full service and add-on measure types. A test of the new
‘div_keyvals_float’ function has also been added.

* A ‘full service’ measure entirely replaces the services of a
comparable baseline technology (e.g., a cold climate heat pump in lieu
of an existing furnace); an ‘add-on’ measure is paired with the
services of a comparable baseline technology (e.g., a window attachment
or BAS controls software for more efficient HVAC operations).