Updates to support and expect start/end years metadata
To be able to examine the AEO data files for the range of years present in each file and depend on the resulting common start and end years, the data import modules and their corresponding tests were modified.

In each data import module, a class was created to contain the filenames for all of the EIA data files imported during the execution of that module. This class is separate from the UsefulVars() class that had already been added to several of these modules to remediate the dependence on global variables.

The data import modules were also modified to depend on the common start and end years to be found in the anticipated metadata JSON file. In some cases, this change required more complicated modifications because some modules had not originally been set up to rely on a single definition for the range of years to be expected from the data.

These changes lay the groundwork for addressing issue #110.