Added module for start/end year metadata determination
Added a module that reads from the newly created EIAData classes in each of the EIA data import modules and scans each of the files named in those classes to determine the earliest and latest years of data reported in each file. Then, the latest start year and earliest end year is identified from among the years found in all of those files. The resulting common start year and end year are output to a small metadata JSON file.

Closes #110.