Automatically add new/updated measures to active measures list
Any measures finalized in ’measures_prep.py’ are now automatically
added to a list of active measure names for further analysis in
‘run.py’ if the measure name is not already on this list. The list is
stored in a ‘active_measnames.json’ file.

In addition to these changes to ‘measure_prep.py,’ the ‘run.py’ routine
required a minor update to accommodate slight restructuring of this
JSON file, which has not yet been posted to the repository.