Initial documentation commit
Initial commit of documentation written using Sphinx, including content from the GitHub wiki, which was the original location of some limited documentation for Scout. Includes updates to the .gitignore file to skip over the rendered reStructuredText file directories. Modified the Sphinx configuration file 'conf.py' from the sphinx-quickstart default to have a more flexible specification for the HTML theme used to render the documentation. The configuration file now uses the build environment and local availability of the Read the Docs theme to determine what theme should be applied to the built HTML documentation. Closes #137.