Add quick start guide and tutorials
Add a quick start guide with instructions on how to set up one's computer to be able to use Scout, including instructions on how to set up Sublime Text for authoring and editing Python and JSON files. Also add an outline of the basic tutorials for users to get started using the major features of Scout. Closes #75. Closes #141.

Rename 'Quick Start Guide' to 'Setup Guide' since the section primarily focuses on how to set up a computer to use Scout.