Minor changes to 'measures_prep.py' and 'run.py'
In ‘measures_prep.py,’ messages printed to the user in the cost
conversion function now report the building type associated with the
conversion in cases where a measure’s installed cost is broken out by
building type. A slight adjustment to the ‘CostConversionTest’ of
‘measures_prep_test.py’ was required to accommodate the new user
message information.

In ‘run.py,’ new data on baseline and efficient energy and carbon costs
have been added to the output JSON.