Expanded cost conversion, handled missing baseline data
The ‘convert_costs’ function in ‘measures_prep.py’ now handles the
following cases:

1) a residential sensors and controls measure where a cost in $/node
must be converted to $/ft^2 floor, and
2) a residential sensors and controls measure where a cost in $/ft^2
floor must be converted to $/unit (all residential technology costs are
in $/unit aside from envelope technologies, which are in $/ft^2 floor).

In case 1, the supporting conversion data (not yet posted to Github)
assume that residential sensor networks require one node per room and
use RECS data to determine the typical number of rooms per residential
building type. In case 2, one technology unit per household is assumed
for all technologies aside from lighting, where households typically
have multiple lighting fixtures. Data on the number of lighting
fixtures of a given class per residential building type are available
in the ‘rsmlgt.txt’ NEMS input file.

Note that in case 1, the conversion listed in case 2 is needed to
finalize the cost ($/node -> $/ft^2 floor -> $/unit).

The ‘CostConversionTest’ of ‘measures_prep_test.py’ has been expanded
to accommodate the above cases. The user message for the
‘convert_costs’ function has also been amended to flag a case where the
conversion was not successful.

Additionally, the following potential gaps in baseline technology data
imported by ‘measures_prep.py’ have been identified and are now handled
in the code:

1) missing baseline technology stock and energy data come in as blank
dicts instead of as previously expected “NA” values,
2) missing baseline technology cost, performance, and/or lifetime data
come in as a dict with year keys but “NA” or 0 values, and
3) missing baseline technology cost, performance, and/or lifetime data
come in as zero values.

In case 1, the baseline microsegment with missing stock and energy data
is removed from the analysis, while in cases 2 and 3, the missing
baseline cost, performance, and/or lifetime values are assumed to be
identical to measure, cost, performance, and lifetime.

Data for a sample ‘lighting gain’ microsegment in the
‘MarketUpdatesTest’ of ‘measures_prep_test.py’ was adjusted to
represent case 3.