Fixed primary/secondary linking, captured fraction weighting
The approach for linking primary and secondary microsegments was
failing for cases where ‘None’ was present in the key chain string used
to link these microsegments (‘None’ occurs for microsegments without a
specific technology type, like residential gas water heating). A
revised approach uses the ‘literal_eval’ function from the ‘ast’ module
to break the key chain string into the climate zone, building type, and
building vintage information needed to make the primary/secondary link.

Additionally, a bug was fixed in the calculation that determines the
total captured fraction to use in adjusting total measure stock,
energy, carbon, and cost figures down following competition with other
measures. This fraction is initialized in the measure’s market entry
year as the portion of the total stock the measure has captured in that
year; in all subsequent years, it represents a weighted combination of
the competed and captured portion of the stock for the current year and
the total captured portion of the stock for all previous years since
market entry. The bug came from the code failing to consistently
initialize the total captured fraction based on the measure’s market
entry year.

Several test output values in the ‘ComCompeteTest’ of ‘run_test.py’
were adjusted to reflect the bug fix.