Change CBECS square footage by vintage data to JSON format
The data used to map EnergyPlus building vintages to Scout building
vintages was previously drawn from a CBECS XLSX file (‘b34.xlsx’). With
this commit, the relevant data from that file have been added to a new
‘CBECS_sf_byvintage.json’ file (not yet posted) in the ‘convert_data’
folder. This change removed the need for the ‘cbecs_vintage_sf’
function in ‘measures_prep.py’.

The ‘EPlusGlobalsTest’ and ‘UpdateMeasuresTest’ in
‘measures_prep_test.py’ were amended to accommodate the above changes.
The ‘CBECS_sf_byvintage.json’ file will be subsequently posted to the
Github repository.