Add section for cost and performance units
Add new section to the documentation that lists the installed cost and energy performance units that should be used when authoring new ECMs. Move the installed cost units information from the analysis approach section and add the performance units information. Update references in the documentation (including those waiting for the newly created performance units details) to point to the new language.