Adjusted handling of supply-demand overlaps and output reporting
When the applicable baseline markets of a supply-side heating/cooling
measure and a demand-side heating/cooling measure overlap, the effect
of each measure’s overlapping savings on the other is reflected as a
reduction in the other measure’s total baseline energy, carbon, and/or
cost. Previously, these effects had been reflected as an increase in
the other measure’s total efficient energy, carbon, and/or cost, which
was inaccurate.

This adjusted handling of supply-demand overlaps required changing many
of the values in the ‘ResCompeteTest’ of ‘run_test.py’ (this test
covers a supply-demand overlap adjustment).

An additional change adds ‘(low)’ and ‘(high)’ variants of total
efficient energy, carbon, and cost outputs to the analysis engine
results JSON.