Improve links to ECM reference section
In sections of the documentation where there is discussion about the particular options available for the applicable baseline market or the units of cost or performance, add links to the approrpiate ECM reference sections or sub-sections. Also replace all of the styled single- and double-quote characters in the Analysis Approach section with plain quote characters.