Supported adding energy/cost benefits for ECM packages
Users may now specify an additional ‘energy savings increase’ and/or
‘cost reduction’ for package ECMs that will be applied to that
package’s energy, carbon, and cost data through a new
‘apply_pkg_benefits’ function. Each benefit is specified as a fraction
that represents a percentage improvement in energy savings or percent
reduction in the overall installed cost of the package ECM.

To support the above change, the ‘MergeMeasuresTest’ in
‘ecm_prep_test.py’ was renamed to ‘MergeMeasuresandApplyBenefits’ and
the test was expanded to cover the new ‘apply_pkg_benefits’ function. A
small change was also required to the ‘AddUncoveredPackagesTest’ test
to reflect the addition of a ‘benefits’ attribute for MeasurePackage
objects.

Closes #149