Added updated conversion data, corrected handling of carbon price
The data for energy and CO2 costs, site to source conversions, and fuel
CO2 intensities have been updated in a new
’site_source_co2_conversions.json’ file that replaces the old
‘cost_s-s_co2.txt’ file. The UsefulVars object in ‘ecm_prep.py’ has
been adjusted to accept data from this new file.

Additionally, the imported CO2 cost data are now multiplied by 1000000
to convert from $/MTon to $/MMTon (Scout CO2 data are in units of
MMTon).

Finally, the energy and CO2 costs, site to source conversions, and fuel
CO2 intensities used in ‘ecm_prep_test.py’ are now hard coded into the
tests based on the previous version of these data (‘cost_s-s_co2.txt’).
This ensures that the tests are not tied to the version of the data
read into ‘ecm_prep.py’ as they were before.

Closes #151