Simplified supply-demand adjustment, adjusted market share calculations
The adjustment of overlaps between supply-side and demand-side heating
and cooling microsegments now follows the following rule:

- If both supply-side and demand-side energy, carbon, and cost
microsegments are present in ECM competition, divide each supply-side
and demand-side microsegment energy, carbon and cost total by two.

The above rule reflects the even apportionment of supply-side and
demand-side energy savings potential in cases where both supply-side
and demand-side heating and cooling ECMs are under consideration. This
approach greatly simplifies the adjustment for supply-demand overlaps,
removing the need for the ‘adj_htcl_add_rec’ function.

Additionally, market share calculations were changed for years in which
none of the competing measures are on the market. In such cases, the
market is now split evenly across the competing measures (previously,
all market shares were zero).

Finally, the calculation that weights total market share for an ECM
between current and previous years now only considers previous years in
which the ECM was on the market in the weighting.

Outputs in ’ResCompeteTest’ and ‘ComCompeteTest’ in ‘run_test.py’ have
been adjusted to reflect the above changes. Note that in these tests,
the ‘test_compete_res_dist’ and ‘test_compete_com_dist’ functions have
been disabled pending changes in the handling of uncertainty in ECM
inputs as part of the alpha testing.