Added XLSX write out and annotations to plotting function
The raw data behind each of the energy, carbon, and cost plots is now
written into a ‘Summary_Data’ XLSX file for each of the adoption
scenarios. The data are organized by year and results scenario (e.g.,
baseline/efficient, competed/uncompeted).

Additionally, the applicable end uses for each individual ECM are now
annotated on the plots, and the 2030 savings number is shown on the
plot covering all ECMs.