Set ECM replacement fraction to zero
This change reflects a simplifying assumption that once an ECM has been
adopted into its baseline market segment, that market segment remains
with the ECM for the rest of the simulation time horizon. Accordingly,
competed market share adjustments need not account for ECM replacements
as part of the competed stock in each year.

Going forward, ECM replacements might be reintroduced to stock-and-flow
and competition calculations such that the model can reflect
improvements in ECM performance over time or account for replacements
of one ECM with another more efficient/cost-effective ECM.

The above adjustment requires revision of several output values in the
MarketUpdatesTest; this test has been commented out for the moment to
allow alpha testing to proceed on this updated code - an issue will be
created.