Update results figure information
Add information and examples to the tutorials about the two new results figures added to plots.R in 1d149acc63d6877d33b15a0fdd7488ffdcaca859 (and a few earlier commits). Also move the Social Cost of Carbon URL reference to conf.py (and thus to the header of every section) so that it is common to all sections of the documentation.