Changes to 'ecm_prep.py,' 'run.py,' and 'plots.R' for beta testing
Multiple changes were made to ‘plots.R,’ ‘ecm_prep.py,’ and ‘run.py’ to
correct various minor errors and omissions in the code and ensure
readiness for beta testing. The key changes are listed below.

‘plots.R’:

-	Minor change to formatting of ‘Multiple’ category ECM points in the
cost effectiveness plot.

‘ecm_prep.py’:

-	Added missing AEO end use names to the ‘Other’ end use classification
for results.
-	Set sub-market scaling fractions to 1 (instead of None) for ECMs
without any sub-market scaling.
-	Added recording and reporting out of the sub-market scaling fraction
associated with each contributing microsegment for each ECM. This
allows sub-market scaling to be accounted for in ECM competition
calculations in ‘run.py’.
-	Added back future ECM replacements to competed stock calculations
(e.g., consider that an ECM can be replaced with itself or another,
more efficient ECM in future years). Previously, future ECM
replacements had caused problems with ECM competition results; however,
these problems were removed with a recent adjustment to the competition
routine.
-	Simplified the routine in ECM package construction that removes
overlaps between the markets of ECMs to be packaged. Previously, this
routine had handled ‘add-on’ and ‘full service’ ECMs differently, when
this was not necessary. The routine now also adjusts sub-market scaling
data for the overlapping ECMs.

‘run.py’:

-	Added missing AEO end use names to the ‘Other’ end use classification
for results.
-	Corrected error in the results end use classification of ECMs with
the ‘other (grid electric)’ AEO end use name.
-	Rounded non-integer ECM/baseline lifetime inputs to the
‘metric_update’ function to the nearest integer. Previously, the floor
of each number was used.
-	Added handling of NaN outcomes in IRR and simple payback calculations.
-	Added a ‘find_added_sbmkt_fracs’ function that apportions the
fraction of a competed market that an ECM does not apply to due to
sub-market scaling across the other competing ECMs.

Minor edits were made to several tests in ‘ecm_prep_test.py’ and
‘run_test.py’ to accommodate the above changes.